Use a fixed size for len
Since the input length is XORed into the final hash, it needs to be a
consistent length to produce consistent output on different platforms.

+++ quickxorhash.c
@@ -13,7 +13,7 @@ struct qxhash {
     size_t    width;
     size_t    shift;
     size_t    shifted;
-    size_t    len;
+    uint64_t  len;
     size_t    cell_len;
     uint64_t *cell;
 };
